+++ plugins/generic/custom/video_src-responsive.js
@@ -5,7 +5,7 @@ module.exports = {
         return {
             href: meta.video_src,
             type: meta.video_type,
-            rel: CONFIG.R.player,
+            rel: [CONFIG.R.player, CONFIG.R.iframely],
             "aspect-ratio": meta.video_width / meta.video_height
         };
     }