+++ WHATSNEW.md
@@ -1,4 +1,21 @@
-0.3.10
+2013.07.19, Version 0.4.0
+
+ + added readtapestry.com plugin
+ + added whosay.com photo plugin
+ + added jsfiddle.net plugin
+ + added theoatmeal.com plugin
+ + added lolwall.co plugin
+ + added openstreetmap.org plugin (Mathias Panzenböck)
+ + added marketplace.firefox.com plugin (Mathias Panzenböck)
+ * fixed smugmug.com plugin
+ + added pinterest.com pin-board widget plugin
+ + added pinterest.com user widget plugin
+ + added pinterest.com pin widget plugin
+ * fixed facebook.video plugin
+ * fixed businessinsider.com plugin
+ * fixed: iframely selects better media sizes for duplicate http/https links from different plugins, giving priority to the ones that produce responsive widget
+
+2013.07.11, Version 0.3.10
 
  * port to htmlparser2, SAX was buggy (Mathias Panzenböck)
  + added /supported-plugins-re.json endpoint
@@ -8,14 +25,14 @@
  * fixed iframed widgets buggy resizing
  * fixed dynamic reader embedding
 
-0.3.9
+2013.07.05, Version 0.3.9
 
  + tindeck.com (Mathias Panzenböck)
  + maps.google.com (Mathias Panzenböck)
  + magnatune.com (Mathias Panzenböck)
  + ocreamix.org (Mathias Panzenböck)
 
-0.3.8
+2013.06.28, Version 0.3.8
 
  + trutv.com
  + video.nhl.com
@@ -26,7 +43,7 @@
  + wikimedia articles and pictures
  + parse links by iframely (oEmbed/2) spec
 
-0.3.7
+2013.06.27, Version 0.3.7
 
  + instagram video
  + 'iframely-widget' styles used by iframely.js to identify generated widgets